requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
python-dotenv
pydrive
PyMuPDF
faiss-cpu
sentence-transformers
